requirements.txt
beautifulsoup4==4.9.3
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.3
lxml==4.6.2
MarkupSafe==1.1.1
PyMySQL==1.0.2
python-dateutil==2.8.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.23
tqdm==4.56.0
urllib3==1.26.3
Werkzeug==1.0.1
